[STO-32] Tests for position calculating

Closes #25

See merge request xheinz/pv260-stock-report!7

+++ DiffCalculator/Model/IndexRecordDiffDto.cs
@@ -8,7 +8,7 @@ public class IndexRecordDiffDto
     public string? Ticker { get; set; }
     public int? DayDiff { get; set; }
     public long? SharesDiff { get; set; }
-    public long? SharesDiffPercentage { get; set; }
+    public double? SharesDiffPercentage { get; set; }
     public double? MarketValueDiff { get; set; }
     public double? WeightDiff { get; set; }
     public bool IsNew { get; set; }
